(feat) Add wifi support
This proved tricky, because most of the docs are for either using esp-wifi with nostd and embassy or esp-idf-svc without embassy.
I had to turn on wifi debugging in the esp-idf layer before this started to work. Not sure if the debugging fixed some timing
issue or maybe the recompile of esp-idf helped. Will have to try again on a clean build to see what happens next.